move function action page
update for WEB env
