+++ features/steps/execure_script.py
@@ -6,6 +6,7 @@ from ManagementElements.Page import Page
 from ManagementElements.Elements import Elements
 from ManagementElements.Locator import Locator
 from Utilities.action_android import ManagementFileAndroid
+from Utilities.common import common_device
 
 dict_yaml = {}
 dict_page = {}
@@ -40,11 +41,12 @@ def click_action(context, element):
     context.element_page = ManagementFile().get_element(context.page_present, element,
                                                         context.device.get_platform_name(), context.dict_save_value)
     if context.device.get_platform_name() == "WEB":
-        ManagementFile().action_page(context.element_page, "click", context.driver, "", context.wait,
-                                     context.dict_save_value, context.device)
+        common_device().action_page(context.element_page, "click", context.driver, "", context.wait,
+                                    context.dict_save_value, context.device)
     elif context.device.get_platform_name() == "ANDROID":
         ManagementFileAndroid().action_page(context.element_page, "click", context.driver, "",
                                             context.dict_save_value, context.device)
+
 
 
 @step(u'I type "{text}" into element {element}')
@@ -52,8 +54,8 @@ def type_action(context, text, element):
     context.element_page = ManagementFile().get_element(context.page_present, element,
                                                         context.device.get_platform_name(), context.dict_save_value)
     if context.device.get_platform_name() == "WEB":
-        ManagementFile().action_page(context.element_page, "type", context.driver, text, context.wait,
-                                     context.dict_save_value, context.device)
+        common_device().action_page(context.element_page, "type", context.driver, text, context.wait,
+                                    context.dict_save_value, context.device)
     elif context.device.get_platform_name() == "ANDROID":
         ManagementFileAndroid().action_page(context.element_page, "type", context.driver, text, context.dict_save_value,
                                             context.device)
@@ -100,8 +102,8 @@ def step_impl(context, element):
     context.element_page = ManagementFile().get_element(context.page_present, element,
                                                         context.device.get_platform_name(), context.dict_save_value)
     if context.device.get_platform_name() == "WEB":
-        ManagementFile().action_page(context.element_page, "clear", context.driver, "", context.wait,
-                                     context.dict_save_value, context.device)
+        common_device().action_page(context.element_page, "clear", context.driver, "", context.wait,
+                                    context.dict_save_value, context.device)
     elif context.device.get_platform_name() == "ANDROID":
         ManagementFileAndroid().action_page(context.element_page, "clear", context.driver, "", context.dict_save_value,
                                             context.device)
